Move sourceSelector
