+++ animeSources/src/main/java/com/talent/animescrapsources/animesources/AllAnimeSource.kt
@@ -3,6 +3,7 @@ package com.talent.animescrapsources.animesources
 import com.talent.animescrap_common.model.AnimeDetails
 import com.talent.animescrap_common.model.AnimeStreamLink
 import com.talent.animescrap_common.model.SimpleAnime
+import com.talent.animescrap_common.source.AnimeSource
 import com.talent.animescrap_common.utils.Utils.getJson
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
