+++ animeSources/src/main/java/com/talent/animescrapsources/animesources/EnimeSource.kt
@@ -3,6 +3,7 @@ package com.talent.animescrapsources.animesources
 import com.talent.animescrap_common.model.AnimeDetails
 import com.talent.animescrap_common.model.AnimeStreamLink
 import com.talent.animescrap_common.model.SimpleAnime
+import com.talent.animescrap_common.source.AnimeSource
 import com.talent.animescrap_common.utils.Utils.getJson
 import com.talent.animescrap_common.utils.Utils.getJsoup
 import kotlinx.coroutines.Dispatchers
