+++ animeSources/src/main/java/com/talent/animescrapsources/animesources/KissKhSource.kt
@@ -4,7 +4,7 @@ import com.talent.animescrap_common.model.AnimeDetails
 import com.talent.animescrap_common.model.AnimeStreamLink
 import com.talent.animescrap_common.model.SimpleAnime
 import com.talent.animescrap_common.utils.Utils.getJson
-import com.talent.animescrapsources.animesources.AnimeSource
+import com.talent.animescrap_common.source.AnimeSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
