+++ animeSources/src/main/java/com/talent/animescrapsources/animesources/YugenSource.kt
@@ -5,7 +5,7 @@ import com.talent.animescrap_common.model.AnimeStreamLink
 import com.talent.animescrap_common.model.SimpleAnime
 import com.talent.animescrap_common.utils.Utils.getJsoup
 import com.talent.animescrap_common.utils.Utils.postJson
-import com.talent.animescrapsources.animesources.AnimeSource
+import com.talent.animescrap_common.source.AnimeSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
