+++ animeSources/src/main/java/com/talent/animescrapsources/animesources/ZoroSource.kt
@@ -5,6 +5,7 @@ import com.google.gson.JsonParser
 import com.talent.animescrap_common.model.AnimeDetails
 import com.talent.animescrap_common.model.AnimeStreamLink
 import com.talent.animescrap_common.model.SimpleAnime
+import com.talent.animescrap_common.source.AnimeSource
 import com.talent.animescrap_common.utils.Utils.getJson
 import com.talent.animescrap_common.utils.Utils.getJsoup
 import kotlinx.coroutines.Dispatchers
